trusty-*-tcwg: add packages
Bring the list of packages installed in the container to the parity with
packages installed in the tcwg schroot images (which we are deprecating in
favor of docker containers).  The only removal is qemu-user-static, which
we don't use (we use qemu-user), since it brings in several hundred megabytes
of dead weight.

Change-Id: Ia03c7dab4e5a78cd35af9a59530655f05ca5b2c4