trusty-*-tcwg: Don't use VOLUME for bind-mounts
Turns out that dockerfile's VOLUME command doesn't support bind-mounts
on purpose (to stay portable).  It only supports scratch mounts and
mounts from other docker containers (--volumes-from).

The ":" in the VOLUME's argument is treated as a normal character, so we
get a bunch of scratch mounts with long strange names.

So leave /tmp scratch mount in the dockerfile and remove all bind-mounts.

Also copy comments on VOLUME and "VOLUME /tmp" to amd64 and i386
dockerfiles.

Change-Id: Ic0644a93a9d7de7c2792b5ae7e56c59455488ccf