*-tcwg-base: Add ability to run QEMU VMs, including using KVM
The harness implementing running QEMU VMs in docker containers is in
toolchain/jenkins-scripts/start-container-qemu.sh.

Rootfs of the container is shared via NFS with the VM.  For NFS to work inside
container NFS kernel module needs to be loaded, which is easiest achieved
by installing nfs-kernel-server package on the host.

For KVM to work container needs to be started in --privileged mode
to provide /dev/kvm inside the container.

Change-Id: I7cba73e350dfbee6e4dde0797eb4bed90fe0c664