LLP: Ensure PIP packages are updated and define settings when running debug
Two problems were found. Our Image is running a out of date version of
Python packages. This is because we have changed LLP code but haven't
updated the pip packages. It is just a case of rebuilding this image but
lets add -U in anyway.

We need to also define settings file when running runserver to ensure
the correct settings file is loaded

Change-Id: Id3d0162698c7d0619cec7ed73fdeb19e7387d9e6