tcwg-dev: Fix hostname of dev container in message
The note at the end of the start.sh script prints the value of the
hostname variable which is never set in the script. Calling the hostname
command is not an option because it prints the info for localhost.
The approach chosen assumes the machine is always reached via ssh when
running this script, in which case it can rely on SSH_CONNECTION to be
set with the server IP. The format of this environment variable is
documented in ssh (1) manual.

Change-Id: I0cd6291792d528584280008dfcd3c6cad23ea619