Use recent clang release for buildbots using clang.
Use a recent clang release rather than the distribution default for clang
buildbots. The libcxx and libcxxabi bot requires a recent release and we
have seen a miscompile with clang 3.8 that is the default on Ubuntu 16.04
that causes the check-clang tests to fail.

For the clang release we've chosen the current release - 1 which is
currently 6.0.0. As we are only selecting one version for all bots the
logic for selecting clang can be simplified a bit.

Change-Id: Ica7e670468dac1586f453475422866c09bfa478b