tcwg-llvmbot: Install and use gcc 7 from PPA
We need to first install software-properties-common, which contains the
add-apt-repository command. Then we can use that to add the toolchain test
builds PPA which contains gcc and g++ 7. Installing the g++-7 package is
enough to get both packages.

We also update run.sh to explicitly invoke gcc-7 and respectively g++-7 in
the ccache wrappers.

Change-Id: I6624cf31b0a7a135264eebf21230a6d0e7633a1b