tcwg-base: Avoid ssh connection drops to ci.linaro.org
We trigger tcwg-benchmark job on ci.linaro.org from tcwg_bmk-build.sh
script, and then wait for up to a few hours for job to finish.  By
default ssh server on ci.linaro.org drops connection after 600s of
inactivity, so we ping it every 300s to avoid that happening.

Change-Id: Id2eeb8edf5293a543a1a7a8a085688b26afa6531