Use apt-get instead of apt
The 'apt' command warns that it does not have a stable CLI, so should be
used with caution in scripts. The 'apt-get' command has the same
functionality, so we should use that instead.

The 'apt' command has also failed with 'std::bad_alloc' when building
the docker image, though I don't know if this is a bug in 'apt' or
memory exhaustion on the build machine.

Change-Id: I3f421668ccc7fe372321dbb15d1baecb2e0ecbcb