Update clang version for libcxx bots
For building and testing libcxx, we need a compiler with support for the
latest C++ standards, so it makes sense to use the most recent build of
clang we can. Clang 11 has recently been released, but there aren't
pre-built binaries for Arm/AArch64 yet, so we need to use 10.0.1 for
now.

Change-Id: I2c7a40a97cbdaeab87c0bbd6a8220210e94bb6ee