tcwg-base: Add rust components to build Python cryptography package
The jipdate package uses the jira module, which uses cryptography.

Previously we would build this from source on Bionic because the
system pip is too old to recognise the "manylinux2014" tag on
the prebuilt AArch64 wheel that is in PyPi.

The manylinux2014 tag is supported as of pip 19.3, Bionic has
9.0.1.
(https://pip.pypa.io/en/stable/news/#id303)

However, the 3.4.1 release of cryptography added rust code
which means we need cargo, rustc and setuptools_rust to build the wheel.

https://github.com/pyca/cryptography/blob/master/CHANGELOG.rst#34---2021-02-07

"cryptography now incorporates Rust code. Users building cryptography themselves
will need to have the Rust toolchain installed. Users who use an officially
produced wheel will not need to make any changes. The minimum supported Rust version is 1.45.0."

I have tried other soloutions to avoid doing this build,
but found none that worked.

1. Installing python3-cryptography from apt

The problem with this is that Bionic's system pip version (9.0.1)
does not see the apt installed crypto as a candidate. This is likely
because it doesn't have the same Python ABI as the pip installed modules
from PyPi.
(or pip is not new enough to understand that they *are* compatible)

This does work on Focal, where we have pip 20.0.2.

2. Installing all dependencies from apt

This does not work because the apt versions are very old and at least
one is known (I ran jipdate and it failed) to be incompatible. Jinja2
is probably the only one I'd bet on working.

apt is jinja2 2.1,  current pip is 2.11.3 (which would probably work)
apt is pyyaml 3.12, current pip is 5.4.1 (known to be incompatible)
apt is jira 1.0.10, current pip is 2.0.0 (likely incompatible)

3. Update pip so it can pick up the apt installed module, or understand
that it can use the manylinux2014 tagged wheel from PyPi.

Combining system installed pip and user installed pip is specifically
warned against in the pip docs.

https://pip.pypa.io/en/stable/installing/

"If you installed Python from a package manager on Linux, you should
always install pip for that Python installation using the same source."

This also means that having a system python with a user installed pip
is inadviseable.

Which leaves us with installing rust so we can build the wheel locally.

Change-Id: Ie5a5dfd4ca512d9af20a7eda4252ee8c80e62a88