tcwg-llvmbot: Support buildkite agents
This adds support for using a buildkite agent instead
of the buildbot worker.

It will be installed alongside buildbot so that we can
keep all the same memory/cpu limits/comiler configs as
buildbot agents have.

Scripts can tell that we want buildkite by checking
what would be the master URL for buildbot. If it is
"buildkite" then we know to use buildkite-agent.

This works because for buildkite we aren't connecting
to many different master servers. Buildkite has one central
controller that takes the token we give it.

This token is passed in the place where the buildbot password
would normally go.

The token is a combination (in buildbot terms) of the
URL and password. So buildkite uses it to auth and direct
us to the right project.

Change-Id: I508254933d2333582e1682e153122e822b6c63a7