tcwg-host: Migrate workaround for dockerhub's pull limits
... from jenkins-scripts/start-container-docker.sh.

Our workaround to dockerhub's pull limits in start-container-docker.sh
works well, but with one exception -- it doesn't properly work when
container is started on a remote host, which we do a lot in
tcwg-benchmark-* jobs.

To solve this we migrate workaround logic to docker wrapper script,
and add a new docker command for it -- "maybepull".

This makes the whole workaround more robust and easier to understand.

Change-Id: I57b3781c6a78b7095adf7c4833362566c2e32973