tcwg-llvmbot/start.sh: Assign cpu sets to libcxx bots
When you assign a cpuset to a container, its processes
can only run on those numbered cores. This means that
inside the container "nproc" returns that number of cores.

Which means that parallel by default tools like lit,
work as expected.

These cores are not "reserved" so other containers
can use them as long as the container assigned to them
is idle.

And we still have a very high priority set for these bots
so they will be able to grab those cores as needed.

https://docs.docker.com/engine/reference/run/#cpuset-constraint

(the fact that it doesn't reserve the cores was tested
locally, I couldn't find any documentation to cite)

LLDB bots remain as they were.

Change-Id: I8528f7c24caa5dfea8af458cbffa48e5c012ef0f