tcwg-llvmbot: Add virtualenv (aka virtualenv for Python2)
Diana removed this from the base image because by default
it was using Python2 and confusing the release script.
Now we install "python3-venv".

This is fine but the test suite expects to be able to call:
virtualenv --python=python3

It turns out that python3's virtualenv is now called venv
and you must call it with: python3 -m venv ...

Since the test suite explicitly asks for python3 to be used,
I think we're safe to add back 'virtualenv' only to the llvmbot
image.

See:
https: //packaging.python.org/guides/installing-using-pip-and-virtual-environments/

Change-Id: I0ea1f7effbc954e870c3da7cee6de1d53c7709f8