tcwg-llvmbot: Rename build user from twcg-buildslave to tcwg-builbot
This aligns with the llvm project's efforts to move toward inclusive
technical language.

tcwg-buildbot will be the first of a few more restricted accounts
used for specific purposes, vs the one that can do everything.
It will probably be the least privleged bot account when all is done.

This changes just the llvmbots to use this new user.

tcwg-buildbot is added to group/passwd and has en empty .ssh folder
in its home data to keep new-user.sh happy. I've also replaced "slave"
with "bot" in the supporting scripts.

Change-Id: I57970972dbac62a979492f75b560fc986adf67e5