tcwg-llvmbot/start.sh: Handle (ignore) options before " -- "
https://review.linaro.org/c/toolchain/jenkins-scripts/+/40984 added
additional options for the other kinds of containers we start but this
script did not handle those so you got this when trying to update bots:

17:17:53 + bash /tmp/tmp.WKijXlFqfs --verbose true --additional_options '' -- linaro/ci-arm64-tcwg-llvmbot-ubuntu:focal normal linaro-clang-aarch64-quick ****
17:17:53 + image=--verbose
17:17:53 + buildmaster=true
17:17:53 + botname=--additional_options
17:17:53 + password=

Ignore anything before " -- " and update the usage to reflect that you
should use " -- " even if you don't have any other options to pass.

Also hook up the usage function which was declared but never used,
return 0 from it since we're only using it from --help not from
some failure point.

Change-Id: I73f4153c9f21b306db1f1938e5ef5ba59f7885f3