tcwg-dev: Always expose /dev/kvm
"Host" containers may not have /dev/kvm exposed to them, therefore assume
that the host OS always has that device node and expose it to the dev
container.

This works because we don't use nested docker service in our "host"
containers. We merely forward docker.sock and docker binary (aka docker
client) in the "host" container.

Change-Id: I68601e50018139b85aae1367c4c89275834b5bab