tcwg-llvmbot: Use clang 14.0.5 as the host compiler
This should fix assert failures on the SVE VLA 2 stage bot:
https://lab.llvm.org/buildbot/#/builders/198/builds/1044

The reproducer for that does not crash on any other machine
so I can only assume that 13.0.1 is miscompiling clang
in the first stage.

A build in a dev container with 14.0.5 did not hit the same
assert.

I filed an upstream issue with what we know about the issue:
https://github.com/llvm/llvm-project/issues/56178

In case someone else finds it. We can find the specific
change that fixed it if it's needed later.

Change-Id: I42a04d208f6e51bc0d463e327eda217b00c06b3f