tcwg-llvmbot: Correct library paths
Since the clang 15 release it has been using "per target" library
paths. Meanig that libc++ is not in /lib it is in /lib/<what clang decided the triple was>.

This is only a problem for one bot which is flang latest clang.
Which, despite the name, distinguishes itself by building against
libcxx not libstdc++.

This should fix that bot and work for armv7 if needed.

Change-Id: I71b7a472722d9ecee123a1918c623226914b8bef