tcwg-dev: Add heuristic to check for KVM support on the host
Asking docker to expose /dev/kvm on a host which doesn't have it causes it
to error out. This is causing trouble on the tk1 boards which don't have
KVM enabled.

The host containers don't have /dev/kvm exposed to them so it's not
possible to directly check whether the device file is available. Therefore
add an heuristic to determine whether the kernel has KVM enabled.

Change-Id: I608bc70d41fd362b5cc6622d500949d490cd47e3