bionic-amd64: upgrade to java11
Jenkins is now supporting only Java 11, which means
that build nodes running Java 8 are unable to run
the remoting agent.

This change upgrades java to 11 where it can.

The Debian stretch distro has been EOL for 6 months
and should be considered to be dropped.

Note: bionic-amd64-android-lava still installs openjdk-8
as it is required by Android10 cts/vts

Signed-off-by: Kelley Spoon <kelley.spoon@linaro.org>
Change-Id: I930a47df3726fed0f49f1955138bd6b3805456da